refactor: docs and structs moving
