+++ src/protocol/xdr/nfs3/file.rs
@@ -6,6 +6,8 @@
 //! - WRITE: Write data to a file (procedure 7)
 //! - COMMIT: Commit asynchronously written data to stable storage (procedure 21)
 //! - LINK: Create a hard link (procedure 15)
+//! - LOOKUP: Look up a file name in a directory (procedure 3)
+//! - CREATE: Create a regular file (procedure 8)
 //!
 //! The structures implement the XDR serialization/deserialization interfaces for
 //! the request arguments and response data of these operations.
@@ -28,11 +30,8 @@ use super::{
 #[allow(non_camel_case_types)]
 #[derive(Debug, Default)]
 pub struct READ3args {
-    /// File handle for the file to be read
     pub file: nfs_fh3,
-    /// Position within the file to begin reading
     pub offset: offset3,
-    /// Number of bytes of data to read
     pub count: count3,
 }
 DeserializeStruct!(READ3args, file, offset, count);
@@ -42,13 +41,9 @@ SerializeStruct!(READ3args, file, offset, count);
 #[allow(non_camel_case_types)]
 #[derive(Debug, Default)]
 pub struct READ3resok {
-    /// File attributes after the operation
     pub file_attributes: post_op_attr,
-    /// Number of bytes actually read
     pub count: count3,
-    /// True if the end of file was reached
     pub eof: bool,
-    /// The data read from the file
     pub data: Vec<u8>,
 }
 DeserializeStruct!(READ3resok, file_attributes, count, eof, data);
@@ -59,11 +54,8 @@ SerializeStruct!(READ3resok, file_attributes, count, eof, data);
 #[allow(non_camel_case_types)]
 #[derive(Debug, Default)]
 pub struct COMMIT3args {
-    /// File handle for the file to commit
     pub file: nfs_fh3,
-    /// Position within the file to start committing
     pub offset: offset3,
-    /// Number of bytes to commit
     pub count: count3,
 }
 DeserializeStruct!(COMMIT3args, file, offset, count);
@@ -73,9 +65,7 @@ SerializeStruct!(COMMIT3args, file, offset, count);
 #[allow(non_camel_case_types)]
 #[derive(Debug, Default)]
 pub struct COMMIT3resok {
-    /// File attributes before and after the operation
     pub file_wcc: wcc_data,
-    /// Write verifier to detect server restarts
     pub verf: writeverf3,
 }
 DeserializeStruct!(COMMIT3resok, file_wcc, verf);
@@ -86,9 +76,7 @@ SerializeStruct!(COMMIT3resok, file_wcc, verf);
 #[allow(non_camel_case_types)]
 #[derive(Debug, Default)]
 pub struct LINK3args {
-    /// File handle for the target file
     pub file: nfs_fh3,
-    /// Directory and name for the new link
     pub link: diropargs3,
 }
 DeserializeStruct!(LINK3args, file, link);
@@ -100,15 +88,9 @@ SerializeStruct!(LINK3args, file, link);
 #[derive(Copy, Clone, Debug, Default, FromPrimitive, ToPrimitive)]
 #[repr(u32)]
 pub enum stable_how {
-    /// Data may be buffered before writing to stable storage
-    /// The server may return before the data is committed to stable storage
     #[default]
     UNSTABLE = 0,
-    /// Data must be committed to stable storage before returning
-    /// Only the data for this request is guaranteed to be committed
     DATA_SYNC = 1,
-    /// All file system data must be committed to stable storage before returning
-    /// This includes the data and all metadata for this request
     FILE_SYNC = 2,
 }
 impl SerializeEnum for stable_how {}
@@ -119,15 +101,10 @@ impl DeserializeEnum for stable_how {}
 #[allow(non_camel_case_types)]
 #[derive(Debug, Default)]
 pub struct WRITE3args {
-    /// File handle for the file to write
     pub file: nfs_fh3,
-    /// Position within the file to begin writing
     pub offset: offset3,
-    /// Number of bytes of data to write
     pub count: count3,
-    /// How to commit the data to storage
     pub stable: stable_how,
-    /// The data to be written
     pub data: Vec<u8>,
 }
 DeserializeStruct!(WRITE3args, file, offset, count, stable, data);
@@ -137,18 +114,16 @@ SerializeStruct!(WRITE3args, file, offset, count, stable, data);
 #[allow(non_camel_case_types)]
 #[derive(Debug, Default)]
 pub struct WRITE3resok {
-    /// File attributes before and after the operation
     pub file_wcc: wcc_data,
-    /// Number of bytes actually written
     pub count: count3,
-    /// How the data was committed to stable storage
     pub committed: stable_how,
-    /// Write verifier to detect server restarts
     pub verf: writeverf3,
 }
 DeserializeStruct!(WRITE3resok, file_wcc, count, committed, verf);
 SerializeStruct!(WRITE3resok, file_wcc, count, committed, verf);
 
+/// Arguments for the LOOKUP procedure (procedure 3) as defined in RFC 1813 section 3.3.3
+/// Used to look up a file name in a directory
 #[allow(non_camel_case_types)]
 #[derive(Debug, Default)]
 pub struct LOOKUP3args {
@@ -161,11 +136,8 @@ DeserializeStruct!(LOOKUP3args, object);
 #[allow(non_camel_case_types)]
 #[derive(Copy, Clone, Debug)]
 pub enum createhow3 {
-    /// Normal file creation - doesn't error if file exists
     UNCHECKED(sattr3),
-    /// Return error if file exists
     GUARDED(sattr3),
-    /// Use exclusive create mechanism (with verifier)
     EXCLUSIVE(createverf3),
 }
 
@@ -178,6 +150,8 @@ impl Default for createhow3 {
 DeserializeTypeEnum!(createhow3; UNCHECKED=0, GUARDED=1, EXCLUSIVE=2);
 SerializeTypeEnum!(createhow3; UNCHECKED=0, GUARDED=1, EXCLUSIVE=2);
 
+/// Arguments for the CREATE procedure (procedure 8) as defined in RFC 1813 section 3.3.8
+/// Used to create a regular file
 #[allow(non_camel_case_types)]
 #[derive(Debug, Default)]
 pub struct CREATE3args {
@@ -186,11 +160,3 @@ pub struct CREATE3args {
 }
 DeserializeStruct!(CREATE3args, dirops, how);
 SerializeStruct!(CREATE3args, dirops, how);
-
-#[allow(non_camel_case_types)]
-#[derive(Debug, Default)]
-pub struct REMOVE3args {
-    pub object: diropargs3,
-}
-SerializeStruct!(REMOVE3args, object);
-DeserializeStruct!(REMOVE3args, object);
